MRD-628 Move DNTR letter content to our codebase
